refactor(ui): Move findOrCreateSplTokenAccount util into hook file
