+++ apps/ui/src/models/index.ts
@@ -1,7 +1,6 @@
 export * from "./amount";
 export * from "./crossEcosystem";
 export * from "./evm";
-export * from "./solana";
 export * from "./swim";
 export * from "./utils";
 export * from "./wallets";
