--- apps/ui/src/models/solana/index.ts
@@ -1,1 +1,0 @@
-export * from "./utils";
